Fix build for all platforms with SNAPSHOT_PROFILE

+++ android/android-emu/android/opengl/OpenglEsPipe.cpp
@@ -87,8 +87,8 @@ public:
             (void)version;
             renderer->load(stream);
 #ifdef SNAPSHOT_PROFILE
-            printf("OpenglEs preload time: %ld ms\n",
-                    (android::base::System::get()->getUnixTimeUs()
+            printf("OpenglEs preload time: %lld ms\n",
+                    (long long)(android::base::System::get()->getUnixTimeUs()
                     - mLoadStartTime) / 1000);
 #endif
         }
@@ -98,8 +98,8 @@ public:
                 renderer->resumeAll();
             }
 #ifdef SNAPSHOT_PROFILE
-            printf("OpenglEs total load time: %ld ms\n",
-                    (android::base::System::get()->getUnixTimeUs()
+            printf("OpenglEs total load time: %lld ms\n",
+                    (long long)(android::base::System::get()->getUnixTimeUs()
                     - mLoadStartTime) / 1000);
 #endif
         }
@@ -123,8 +123,8 @@ public:
                 renderer->resumeAll();
             }
 #ifdef SNAPSHOT_PROFILE
-            printf("OpenglEs total save time: %ld ms\n",
-                    (android::base::System::get()->getUnixTimeUs()
+            printf("OpenglEs total save time: %lld ms\n",
+                    (long long)(android::base::System::get()->getUnixTimeUs()
                     - mSaveStartTime) / 1000);
 #endif
         }
